services/aqi_agent/pyproject.toml
[project]
name = "aqi-agent"
version = "1.0.0"
description = "Air Quality Index question answering service using Text2SQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgi-correlation-id>=4.2.0",
    "base",
    "easydict>=1.13",
    "fastapi>=0.109.1",
    "langgraph>=0.3.5",
    "langgraph-checkpoint-redis>=0.0.5",
    "lite-llm",
    "logger",
    "opensearch",
    "pg",
    "pydantic-settings>=2.9.1",
    "rapidfuzz>=3.14.3",
    "redis>=5.0.0",
    "simpleeval>=1.0.3",
    "sqlglot>=29.0.1",
    "sqlparse>=0.4.4",
    "structlog>=25.0.0",
    "uvicorn[standard]>=0.23.0",
]

[project.scripts]
aqi_agent = "aqi_agent:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
]

[build-system]
requires = ["uv_build>=0.9.14,<0.11.0"]
build-backend = "uv_build"

[tool.uv.sources]
base = { workspace = true }
lite-llm = { workspace = true }
logger = { workspace = true }
pg = { workspace = true }
opensearch = { workspace = true }
